# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: DLLM LAN API
  version: 0.1.0
  description: Coordinator LAN API. All inference stays on-LAN. SSE resume via Last-Event-ID.
servers:
  - url: https://192.168.1.10:8443
    description: Coordinator (mDNS _dllm._tcp.local.)
paths:
...
  /v1/devices/{id}:
    get:
      operationId: deviceDetail
      summary: Device detail — registry row + honest load + self-only sessions + stage (404 unknown)
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceDetail'
        '404':
          description: unknown device (honest absent, nothing invented)
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean }
                  error: { type: string }
    delete:
      operationId: deleteDevice
      summary: Hard-delete a device row — removes drill junk / stale dupes from the registry (404 unknown; 400 self coordinator, never orphan the mesh). A later heartbeat recreates the row as paired.
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: deleted
          content:
            application/json:
              schema:
                type: object
                required: [ok, id]
                properties:
                  ok: { type: boolean }
                  id: { type: string }
        '400':
          description: refused to delete self coordinator
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean }
                  error: { type: string }
        '404':
          description: unknown device (honest absent, nothing invented)
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean }
                  error: { type: string }
...
components:
  parameters:
...
    DeviceId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    DeviceDetail:
      type: object
      required: [device, load, sessions, stage]
      properties:
        device:
          type: object
          required: [device_id, role, permissions, status, active, last_seen, paired_at, paired_by, cert_fp]
          properties:
            device_id: { type: string }
            device_name: { type: string, nullable: true }
            role: { type: string }
            permissions: { type: array, items: { type: string } }
            status: { type: string, enum: [paired, revoked] }
            active: { type: boolean }
            last_seen: { type: string, format: date-time }
            paired_at: { type: string, format: date-time }
            paired_by: { type: string }
            cert_fp: { type: string }
            capabilities: { type: object, nullable: true }
        load:
          type: object
          required: [source]
          description: live = fresh sysinfo for self; reported = peer row + updated_at; none = no data (never synthesized).
          properties:
            source: { type: string, enum: [live, reported, none] }
            cpu_pct: { type: number, minimum: 0, maximum: 100 }
            mem_pct: { type: number, minimum: 0, maximum: 100 }
            updated_at: { type: string, format: date-time }
        sessions:
          type: array
          description: All current sessions when id == self (single-device serves all); [] for peers.
          items:
            type: object
            required: [id, model, tokens_out]
            properties:
              id: { type: string }
              model: { type: string }
              tokens_out: { type: integer }
              last_token_at: { type: string, format: date-time, nullable: true }
        stage:
          type: object
          nullable: true
          description: Self runs layers 0-27; null for peers.
          properties:
            layer_start: { type: integer, example: 0 }
            layer_end: { type: integer, example: 27 }
...